refactor(virtual): deduplicate virtual output code and add headless dmabuf import
- Extract VirtualOutputs struct into backend/mod.rs shared by Headless
  and Tty, removing ~90 lines of duplication
- Implement Headless::import_dmabuf using the existing GlesRenderer
  instead of unimplemented!(), enabling PipeWire/screencast in
  headless mode
